Task/basic-column-filter-component-wip (#1677)
* The filterclause component opens up the suggestions dropdown by default. In the scenario where we place multiple FilterClauses (without a filtereditor) on the page, they render with open dropdowns.
This dropdown behaviour can now be controlled via a prop.

Also, for multiselect setup, if the valies are cleared, the model will trigger a commit on close of the dropdown. (useFilterClause.ts -> handleOpenChange).
This will enable the consumer to attach a handler based on the value of the filter.

* - Added openDropdownOnFocus prop to the FilterClause component which will help (along with the filterTouched ref) in identifying if the dropdown needs to be shown on initial render

* WIP

* Removing console.info

* Changes made as per PR feedback

* Updated FilterClause.examples.tsx.

* Initial column filter component changes
We can remove the existing ColumnFilter component and rename this new component "BasicColumnFilter" to "ColumnFilter" once we are happy with the implementation.
Will add cypress tests as well once PR feedback is received

* Made changes as per PR feedback
Made the display of search icon prop-driven
Added cypress tests